chore(python): move agent execution logic to executor
